setup: move functionality for AutoProcess 'setup' method into 'commands' module.
