+++ auto_process_ngs/auto_processor.py
@@ -100,6 +100,7 @@ def add_command(name,f):
 # Classes
 #######################################################################
 
+@add_command("setup",commands.setup)
 @add_command("make_fastqs",commands.make_fastqs)
 @add_command("analyse_barcodes",commands.analyse_barcodes)
 @add_command("merge_fastq_dirs",commands.merge_fastq_dirs)
@@ -784,194 +785,6 @@ class AutoProcess(object):
                 return self.load_illumina_data().paired_end
             except IlluminaData.IlluminaDataError:
                 return None
-
-    def setup(self,data_dir,analysis_dir=None,sample_sheet=None):
-        """
-        Set up the initial analysis directory
-
-        This does all the initialisation of the analysis directory
-        and processing parameters
-
-        Arguments:
-          data_dir (str): source data directory
-          analysis_dir (str): corresponding analysis directory
-          sample_sheet (str): name and location of non-default sample
-            sheet file
-
-        """
-        data_dir = data_dir.rstrip(os.sep)
-        if not fileops.exists(data_dir):
-            raise Exception("Data directory '%s' not found" %
-                            data_dir)
-        if not fileops.Location(data_dir).is_remote:
-            data_dir = os.path.abspath(data_dir)
-        run_name = os.path.basename(data_dir)
-        if analysis_dir is None:
-            analysis_dir = os.path.join(
-                os.getcwd(),run_name)+'_analysis'
-        else:
-            analysis_dir = os.path.abspath(analysis_dir)
-        # Create the analysis directory structure
-        if not os.path.exists(analysis_dir):
-            # Make a temporary analysis dir
-            tmp_analysis_dir = os.path.join(
-                os.path.dirname(analysis_dir),
-                ".%s.%s" % (os.path.basename(analysis_dir),
-                            uuid.uuid4()))
-            self.analysis_dir = tmp_analysis_dir
-            logging.debug("Creating temp directory '%s'" %
-                          self.analysis_dir)
-            # Create directory structure
-            self.create_directory(self.analysis_dir)
-            self.log_dir
-            self.script_code_dir
-        else:
-            # Directory already exists
-            logging.warning("Analysis directory '%s' already exists" %
-                            analysis_dir)
-            self.analysis_dir = analysis_dir
-            # check for parameter file
-            if self.has_parameter_file:
-                self.load_parameters()
-            else:
-                logging.warning("No parameter file found in %s" %
-                                self.analysis_dir)
-        # Run datestamp, instrument name and instrument run number
-        try:
-            datestamp,instrument,run_number,flow_cell_prefix,flow_cell_id = \
-                IlluminaData.split_run_name_full(run_name)
-            run_number = run_number.lstrip('0')
-            flow_cell = flow_cell_prefix + flow_cell_id
-        except Exception as ex:
-            logging.warning("Unable to extract information from run name '%s'" \
-                            % run_name)
-            logging.warning("Exception: %s" % ex)
-            datestamp = None
-            instrument= None
-            run_number = None
-            flow_cell = None
-        # Identify missing data and attempt to acquire
-        # Sequencing platform
-        platform = self.metadata.platform
-        if platform is None:
-            platform = bcl2fastq_utils.get_sequencer_platform(
-                data_dir,
-                instrument=instrument,
-                settings=self.settings)
-        print "Platform identified as '%s'" % platform
-        # Log dir
-        self.set_log_dir(self.get_log_subdir('setup'))
-        # Custom SampleSheet.csv file
-        custom_sample_sheet = self.params.sample_sheet
-        if custom_sample_sheet is None:
-            print "Acquiring sample sheet..."
-            if sample_sheet is None:
-                targets = ('Data/Intensities/BaseCalls/SampleSheet.csv',
-                           'SampleSheet.csv',)
-            else:
-                targets = (sample_sheet,)
-            # Try each possibility until one sticks
-            for target in targets:
-                target = fileops.Location(target)
-                tmp_sample_sheet = os.path.join(self.tmp_dir,
-                                                os.path.basename(target.path))
-                if target.is_url:
-                    # Try fetching samplesheet from URL
-                    print "Trying '%s'" % target.url
-                    try:
-                        urlfp = urllib2.urlopen(target.url)
-                        with open(tmp_sample_sheet,'w') as fp:
-                            fp.write(urlfp.read())
-                    except Exception as ex:
-                        logging.warning("Error fetching sample sheet data "
-                                        "from '%s': %s" % (target.url,ex))
-                        tmp_sample_sheet = None
-                else:
-                    # Assume target samplesheet is a file on a local
-                    # or remote server
-                    if target.is_remote:
-                        sample_sheet = str(target)
-                    else:
-                        if os.path.isabs(target.path):
-                            sample_sheet = target.path
-                        else:
-                            sample_sheet = os.path.join(data_dir,
-                                                        target.path)
-                    print "Trying '%s'" % sample_sheet
-                    rsync = applications.general.rsync(sample_sheet,
-                                                       self.tmp_dir)
-                    print "%s" % rsync
-                    status = rsync.run_subprocess(log=self.log_path('rsync.sample_sheet.log'))
-                    if status != 0:
-                        logging.warning("Failed to fetch sample sheet '%s'"
-                                        % sample_sheet)
-                        tmp_sample_sheet = None
-                    else:
-                        break
-            # Bail out if no sample sheet was acquired
-            if tmp_sample_sheet is None:
-                shutil.rmtree(self.analysis_dir)
-                self.analysis_dir = None
-                raise Exception("Unable to acquire sample sheet")
-            # Keep a copy of the original sample sheet
-            original_sample_sheet = os.path.join(self.analysis_dir,
-                                                 'SampleSheet.orig.csv')
-            print "Copying original sample sheet to %s" % original_sample_sheet
-            shutil.copyfile(tmp_sample_sheet,original_sample_sheet)
-            # Set the permissions for the original SampleSheet
-            os.chmod(original_sample_sheet,0664)
-            # Process acquired sample sheet
-            custom_sample_sheet = os.path.join(self.analysis_dir,
-                                               'custom_SampleSheet.csv')
-            sample_sheet = bcl2fastq_utils.make_custom_sample_sheet(
-                tmp_sample_sheet,custom_sample_sheet)
-        else:
-            sample_sheet = IlluminaData.SampleSheet(custom_sample_sheet)
-            original_sample_sheet = os.path.join(self.analysis_dir,
-                                                 'SampleSheet.orig.csv')
-        print "Sample sheet '%s'" % custom_sample_sheet
-        # Library Prep Kit/Assay
-        assay = None
-        for item in ('Assay','Library Prep Kit'):
-            try:
-                assay = IlluminaData.SampleSheet(original_sample_sheet).header[item]
-                break
-            except KeyError:
-                logging.warning("No element '%s' found in sample sheet"
-                                % item)
-        # Bases mask
-        print "Bases mask set to 'auto' (will be determined at run time)"
-        bases_mask = "auto"
-        # Generate and print predicted outputs and warnings
-        print samplesheet_utils.predict_outputs(sample_sheet=sample_sheet)
-        samplesheet_utils.check_and_warn(sample_sheet=sample_sheet)
-        # Move analysis dir to final location if necessary
-        if self.analysis_dir != analysis_dir:
-            logging.debug("Moving %s to final directory" % self.analysis_dir)
-            os.rename(self.analysis_dir,analysis_dir)
-            self.analysis_dir = analysis_dir
-            # Update the custom sample sheet path
-            custom_sample_sheet = os.path.join(
-                analysis_dir,
-                os.path.basename(custom_sample_sheet))
-            print "Created analysis directory '%s'" % self.analysis_dir
-        # Store the parameters
-        self.params['data_dir'] = data_dir
-        self.params['analysis_dir'] = self.analysis_dir
-        self.params['sample_sheet'] = custom_sample_sheet
-        self.params['bases_mask'] = bases_mask
-        self.params['acquired_primary_data'] = False
-        # Store the metadata
-        self.metadata['run_name'] = self.run_name
-        self.metadata['platform'] = platform
-        self.metadata['instrument_name'] = instrument
-        self.metadata['instrument_datestamp'] = datestamp
-        self.metadata['instrument_run_number'] = run_number
-        self.metadata['instrument_flow_cell_id'] = flow_cell
-        self.metadata['assay'] = assay
-        # Set flags to allow parameters etc to be saved back
-        self._save_params = True
-        self._save_metadata = True
 
     def setup_from_fastq_dir(self,analysis_dir,fastq_dir):
         # Do setup for an existing directory containing fastq files
