+++ auto_process_ngs/commands/__init__.py
@@ -7,6 +7,7 @@
 # Imports
 #######################################################################
 
+from setup_cmd import setup
 from setup_analysis_dirs_cmd import setup_analysis_dirs
 from make_fastqs_cmd import make_fastqs
 from analyse_barcodes_cmd import analyse_barcodes
